% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acr2cdai}
\alias{acr2cdai}
\title{Change in CDAI by 6-month ACR Response}
\format{A list of one data.tables with four columns. Column 1 is ACR response category,
 column 2 is the point estimate for the mean change in CDAI at 6 months, and columns 3/4 are 
 lower/upper bounds on the mean change in CDAI. Current default is to set bounds to +/-20% of
 the mean. Format is a list to facilitate inclusion of evidence from additional data sources 
 with different patient populations.}
\source{
Aletaha, Daniel, and Josef S. Smolen. "The simplified disease activity index (SDAI) and 
clinical disease activity index (CDAI) to monitor patients in standard clinical care." 
Best practice & research Clinical rheumatology 21.4 (2007): 663-675.

Aletaha, Daniel, et al. "Acute phase reactants add little to composite disease activity indices
for rheumatoid arthritis: validation of a clinical activity score." Arthritis research & 
therapy 7.4 (2005): R796.
}
\usage{
acr2cdai
}
\description{
Mean of change in CDAI by 6-month ACR response category (ACR < 20, ACR20, ACR50, ACR70).
}
\keyword{datasets}
