% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acr2haq}
\alias{acr2haq}
\title{HAQ change by 6-month ACR response}
\format{A data.table with three columns. First column is ACR response category,
second column is mean change in HAQ at 6 months, and  third column is the standard error
of the mean change in HAQ.}
\source{
Carlson, Josh J., et al. "Economic evaluation of tocilizumab monotherapy compared to adalimumab monotherapy 
in the treatment of severe active rheumatoid arthritis." Value in Health 18.2 (2015): 173-179.
}
\usage{
acr2haq
}
\description{
Mean and standard deviation of HAQ change by mutually exclusive 6-month ACR response category 
(ACR < 20, ACR 20 - <50, ACR 50 - <70, and ACR 70+).
}
\keyword{datasets}
