% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acr_nma2prob}
\alias{acr_nma2prob}
\title{NMA parameters to ACR response probabilities}
\usage{
acr_nma2prob(A, z2, z3, delta)
}
\arguments{
\item{A}{A vector from the posterior distribution of the probability that ACR response < 20 
from cDMARDS.}

\item{z2}{A vector from the posterior distribution for the ACR 50 cutpoint.}

\item{z3}{A vector from the posterior distribution for the ACR 70 cutpoint.}

\item{delta}{A matrix from the posterior distribution of regression coefficient means 
(on probit scale) for each therapy in the NMA.}
}
\description{
Calculate ACR response probabilities from NMA ordered probit parameters.
}
