% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acr_prob}
\alias{acr_prob}
\title{ACR response from NMA}
\usage{
acr_prob(A, z2, z3, delta)
}
\arguments{
\item{A}{Mean probability that ACR response < 20 from cDMARDS}

\item{z2}{Mean for ACR 50 cutpoint}

\item{z3}{Mean for ACR 70 cutpoint}

\item{delta}{Vector of regression coefficient means (on probit scale) for therapies}
}
\description{
Calculate ACR response probability from ordered probit NMA
}
