% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{adjust_utility_wailoo}
\alias{adjust_utility_wailoo}
\title{The IVI-RA simulation}
\usage{
adjust_utility_wailoo(pars)
}
\arguments{
\item{pars}{\code{utility.mixture} object from \code{sample_pars}.}
}
\description{
Run the IVI-RA individual patient simulation model.
}
\keyword{internal}
