% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{apply_summary}
\alias{apply_summary}
\title{Calculate mean, SD, and quantiles}
\usage{
apply_summary(x)
}
\value{
matrix
}
\description{
Calculate mean, SD, and quantiels of matrix or vector
}
\keyword{internal}
