% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_lifetable}
\alias{check_lifetable}
\title{Check lifetable}
\usage{
check_lifetable(lt)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check argument that passes a lifetable is correct.
}
\keyword{internal}
