% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_matrix}
\alias{check_matrix}
\title{Check matrix}
\usage{
check_matrix(m, nrow, ncol, pos = FALSE)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check argument that must be a matrix is correct.
}
\keyword{internal}
