% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nma_k}
\alias{check_nma_k}
\title{Check NMA effect reduction}
\usage{
check_nma_k(lower, upper)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check argument that are passed to \link{sample_pars} related to the reduction
in treatment effects for bDMARD experienced patients.
}
\keyword{internal}
