% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_prod_loss}
\alias{check_prod_loss}
\title{Check parameters of prod_loss}
\usage{
check_prod_loss(simhaq, pl_haq)
}
\arguments{
\item{pl_haq}{\code{pl_haq} as passed to \link{prod_loss}}
}
\description{
Error messages when incorrect inputs are passed to prod_loss.
}
\keyword{internal}
