% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_scalar}
\alias{check_scalar}
\title{Check scalar}
\usage{
check_scalar(s, pos = FALSE)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check argument that must be a scalar is correct.
}
\keyword{internal}
