% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_sim_hc_cost}
\alias{check_sim_hc_cost}
\title{Check parameters of sim_hc_cost}
\usage{
check_sim_hc_cost(simhaq, weight, pars)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}.}

\item{weight}{Weight of patient.}

\item{pars}{\code{pars} as passed to \link{sim_hc_cost}.}
}
\description{
Error messages when incorrect inputs are passed to sim_hc_cost.
}
\keyword{internal}
