% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_sim_qalys}
\alias{check_sim_qalys}
\title{Check parameters of sim_qalys}
\usage{
check_sim_qalys(simhaq, utility, si_ul, x_attr, tx_attr_ug)
}
\arguments{
\item{si_ul}{\code{si_ul} as passed to \link{sim_qalys}}
}
\description{
Error messages when incorrect inputs are passed to sim_qalys
}
\keyword{internal}
