% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_sim_utility_mixture}
\alias{check_sim_utility_mixture}
\title{Check parameters of sim_utility_mixture}
\usage{
check_sim_utility_mixture(simhaq, male, pars)
}
\arguments{
\item{male}{Vector indiciating patient gender (1 = male, 0 = female)}

\item{pars}{\code{pars} as passed to \link{sim_utility_mixture}}
}
\description{
Error messages when incorrect inputs are passed to sim_utility_mixture.
}
\keyword{internal}
