% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_sim_utility_wailoo}
\alias{check_sim_utility_wailoo}
\title{Check parameters of sim_utility_wailoo}
\usage{
check_sim_utility_wailoo(simhaq, input_data, pars)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}}

\item{pars}{\code{pars} as passed to \link{sim_utility_mixture}}

\item{male}{Vector indiciating patient gender (1 = male, 0 = female)}
}
\description{
Error messages when incorrect inputs are passed to sim_utility_wailoo.
}
\keyword{internal}
