% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ttd}
\alias{check_ttd}
\title{Check time to discontinuation}
\usage{
check_ttd(pars, x_ttd)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check that time to discontinuation argument in \link{sample_pars} is correct.
}
\keyword{internal}
