% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_vector}
\alias{check_vector}
\title{Check vector}
\usage{
check_vector(v, len = NULL, pos = FALSE)
}
\value{
Error message if incorrect arguments are passed.
}
\description{
Check argument that must be a vector is correct.
}
\keyword{internal}
