% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eular2haq}
\alias{eular2haq}
\title{HAQ change by 6-month Eular Response}
\format{A matrix with rows as Eular response categories. First column is mean and second 
column is standard deviation.}
\source{
Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, 
golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously
 treated with disease-modifying antirheumatic drugs and after the failure of conventional 
 disease-modifying antirheumatic drugs only: systematic review and economic evaluation." 
Health Technology Assessment 20.35 (2016): 1-610.
}
\usage{
eular2haq
}
\description{
HAQ change by 6-month Eular response
}
\keyword{datasets}
