% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haq.lcgm.pars}
\alias{haq.lcgm.pars}
\title{Parameters for Latent Class Growth Model (LCGM)}
\format{List containing coefficient vector, variance-covariance matrix, and indices for 
parameters.}
\source{
Norton, Sam, et al. "Health Assessment Questionnaire disability progression in early
 rheumatoid arthritis: systematic review and analysis of two inception cohorts." 
 Seminars in arthritis and rheumatism. Vol. 44. No. 2. WB Saunders, 2014.
}
\usage{
haq.lcgm.pars
}
\description{
Parameters for Norton (2014) LCGM used to simulate non-linear HAQ trajectories for
4 latent classes.
}
\keyword{datasets}
