% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haq.lprog.age}
\alias{haq.lprog.age}
\title{Linear HAQ progression rate differences by age}
\format{A matrix with rows as age category. First column is mean and second 
column is standard error}
\source{
Michaud, Kaleb, Gene Wallenstein, and Frederick Wolfe. "Treatment and nontreatment
 predictors of health assessment questionnaire disability progression in 
 rheumatoid arthritis: a longitudinal study of 18,485 patients." 
 Arthritis care & research 63.3 (2011): 366-372.
}
\usage{
haq.lprog.age
}
\description{
Impact of age on annual linear HAQ progression rate.
}
\keyword{datasets}
