% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp.cost}
\alias{hosp.cost}
\title{Hospitalization costs}
\format{A data table with 5 variables.
\describe{
  \item{haq}{HAQ score range.}
  \item{days_mean}{Mean number of days in the hospital for a given HAQ score range.}
  \item{days_se}{Standard error of number of days in the hospital for a given HAQ score range.}
  \item{cost_pday_mean}{Mean number of costs per day in the hospital for a given HAQ score range.}
  \item{cost_pday_se}{Standard error of costs per day in the hospital for a given HAQ score range.}
}}
\source{
Carlson, Josh J., et al. "Economic evaluation of tocilizumab monotherapy compared to adalimumab monotherapy 
in the treatment of severe active rheumatoid arthritis." Value in Health 18.2 (2015): 173-179.
}
\usage{
hosp.cost
}
\description{
Hospitalization costs due to RA.
}
\keyword{datasets}
