% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence}
\alias{incidence}
\alias{incidence.female}
\alias{incidence.male}
\title{Incidence}
\format{A data frame with 101 rows and 4 variables:
\describe{
  \item{age}{Age in years.}
  \item{events}{Number of events (i.e., number of patients with RA).}
  \item{person_years}{Person years, or, time at risk of event.}
  \item{incidence_rate}{Annual incidence rate.}

}}
\source{
Myasoedova, Elena, et al. "Is the incidence of rheumatoid arthritis rising?: 
results from Olmsted County, Minnesota, 1955–2007." 
Arthritis & Rheumatology 62.6 (2010): 1576-1582.
}
\usage{
incidence.female

incidence.male
}
\description{
Annual incidence rate of rheumatoid arthritis in the United States.
}
\keyword{datasets}
