% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi-ra.R
\docType{package}
\name{iviRA}
\alias{iviRA}
\alias{iviRA-package}
\title{iviRA: an R package for running the IVI-RA model}
\description{
An R package for running the IVI-RA model, an individual patient simulation that
simulates the costs, health outcomes, and risks associated with disease-modifying anti-rheumatic 
drugs (DMARDs) and biologic DMARDSs (bDMARDs) for patients with moderate to severe 
rheumatoid arthritis.
}
\section{Resources}{

 \itemize{
 \item Package website (\url{https://innovationvalueinitiative.github.io/IVI-RA}) 
 \item Soure code (\url{https://github.com/InnovationValueInitiative/IVI-RA}) 
 \item Report bugs (\url{https://github.com/InnovationValueInitiative/IVI-RA/issues})
 }
}

\author{
\strong{Maintainer}: Devin Incerti  \email{devin.incerti@thevalueinitiative.org} \cr 

Contributors: Jeroen Jansen \email{jeroen.jansen@thevalueinitiative.org}
}
