% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{lt_data}
\alias{lt_data}
\title{Lifetable for simulation}
\usage{
lt_data(ltfemale, ltmale)
}
\arguments{
\item{ltfemale}{}
}
\value{
A list containing lifetables for females and males with qx transformed using the logit function. 
The lifetable has three columns: age, qx, and logit_qx.
}
\description{
Generate lifetable matrix for use in simulation
}
