% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{lt_pars}
\alias{lt_pars}
\title{Lifetable for simulation}
\usage{
lt_pars(ltfemale, ltmale)
}
\arguments{
\item{ltfemale}{Lifetable for women. Must contain column 'age' for single-year of age and 'qx' for
the probability of death at a given age. Age must range from 0 to 100.}

\item{ltmale}{Identical to \code{ltfemale} but for men.}
}
\value{
A list containing lifetables for females and males with qx transformed using the logit function. 
The lifetable has three columns: age, qx, and logit_qx.
}
\description{
Generate lifetable matrix for use in simulation
}
