% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgmt.cost}
\alias{mgmt.cost}
\title{Cost of general management of RA}
\format{A data table with 4 variables.
\describe{
  \item{service}{Medical service performed}
  \item{est}{Point estimate for cost}
  \item{lower}{Lower range for cost}
  \item{upper}{Upper range for cost}
}}
\source{
{Claxton, Lindsay, et al. "An Economic Evaluation of Tofacitinib Treatment in Rheumatoid 
Arthritis: Modeling the Cost of Treatment Strategies in the United States." Journal of managed
care & specialty pharmacy 22.9 (2016): 1088-1102.}
}
\usage{
mgmt.cost
}
\description{
Cost of general management of RA.
}
\keyword{datasets}
