% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-prob.R
\name{mlogit_prob}
\alias{mlogit_prob}
\title{Predicted probabilities for multinomial logistic regression}
\usage{
mlogit_prob(x, beta)
}
\arguments{
\item{x}{Row vector containing variables in the model for a single individual.}

\item{beta}{Matrix of coefficients; one for each outcome other than the reference outcome.}
}
\value{
Vector of predicted probabilities for the individual.
}
\description{
Predicted probabilitities from a multinomial logistic regression for a single individual. This function
is currently not vectorized.
}
