% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_defaults}
\alias{modify_defaults}
\title{Modify argument defaults}
\usage{
modify_defaults(default_args, control_args)
}
\description{
Modify argument default values generated from \code{formals} with user controlled argument values.
}
