% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{mom_gamma}
\alias{mom_gamma}
\title{Method of moments for Gamma Distribution}
\usage{
mom_gamma(mean, sd, quant = NULL)
}
\value{
parameters of gamma distribution. Optionally return quantiles of
distribution as well.
}
\description{
Method of moments for Gamma Distribution
}
