% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nma}
\alias{nma}
\alias{nma.acr.naive}
\alias{nma.das28.naive}
\alias{nma.haq.naive}
\title{NMA estimates}
\format{Each dataset is a list with two elements
\describe{
  \item{mean}{Mean of the parameters.}
  \item{vcov}{Variance-covariance matrix of the parameters.}
}}
\usage{
nma.acr.naive

nma.das28.naive

nma.haq.naive
}
\description{
Estimates of ACR response, change in DAS28, and change in HAQ at 6 months for 
bDMARD naive patients from Bayesian random effects NMAs.
}
\details{
ACR response parameters were estimated using an ordered probit model. Parameters for
DAS28 and HAQ were estimated using models for continuous data with a normal likelihood and 
identity link. The mean and variance-covariance matrix were estimated using the sample means
and sample covariance matrices from the posterior distributions of the Bayesian analyses.
}
\keyword{datasets}
