% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nma_acrprob}
\alias{nma_acrprob}
\title{NMA parameters to ACR response probabilities}
\usage{
nma_acrprob(A, z2, z3, d, rr = 1)
}
\arguments{
\item{A}{A vector from the posterior distribution of the probability that ACR response < 20 
from cDMARDS.}

\item{z2}{A vector from the posterior distribution for the ACR 50 cutpoint.}

\item{z3}{A vector from the posterior distribution for the ACR 70 cutpoint.}

\item{d}{A matrix from the posterior distribution of the "d's" 
for each treatment in the NMA.}

\item{rr}{Sampled value of elative risk.}
}
\value{
List containing an array \code{non.overlap} (the probability of ACR < 20, ACR 20-50,
ACR 50-70, and ACR 70+) and \code{overlap} (the probability of ACR20/50/70).
}
\description{
Calculate ACR response probabilities from NMA ordered probit parameters.
}
