% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nma_lm2prob}
\alias{nma_lm2prob}
\title{NMA parameters to change in continuous outcome}
\usage{
nma_lm2prob(A, delta)
}
\arguments{
\item{A}{A vector from the posterior distribution of the probability that ACR response < 20 
from cDMARDS.}

\item{delta}{A matrix from the posterior distribution of regression coefficients 
for each therapy in the NMA.}
}
\value{
Matrix of change in continuous outcome variable for each therapy.
}
\description{
Calculate change in continuous outcome from linear model used for NMA.
}
