% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nma_lm_dy}
\alias{nma_lm_dy}
\title{NMA parameters to change in continuous outcome}
\usage{
nma_lm_dy(A, d)
}
\arguments{
\item{A}{A vector from the posterior distribution of the probability that ACR response < 20 
from cDMARDS.}

\item{d}{A matrix from the posterior distribution of the "d's" 
for each treatment in the NMA.}
}
\value{
Matrix of change in continuous outcome variable for each therapy.
}
\description{
Calculate change in continuous outcome from linear model used for NMA.
}
