% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-prob.R
\name{ologit_prob}
\alias{ologit_prob}
\title{Predicted probabilities for ordered logistic regression}
\usage{
ologit_prob(x, beta, cut)
}
\arguments{
\item{x}{Row vector containing variables in the model for a single individual.}

\item{beta}{Vector of coefficients.}

\item{cut}{Vector of cutpoints, one for each category.}
}
\value{
Vector of predicted probabilities for the individual.
}
\description{
Predicted probabilitities from an ordered logistic regression for a single individual. This function
is currently not vectorized.
}
