% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pain}
\alias{pain}
\title{Correlation between HAQ and pain}
\format{List containing mean and variance of HAQ and pain scores, as well as the 
correlation between pain and HAQ.}
\source{
Sarzi-Puttini, Piercarlo, et al. "Correlation of the score for subjective pain with
 physical disability, clinical and radiographic scores in recent onset rheumatoid arthritis."
 BMC musculoskeletal disorders 3.1 (2002): 1..
}
\usage{
pain
}
\description{
Correlation between HAQ and pain
}
\keyword{datasets}
