% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prod.loss}
\alias{prod.loss}
\title{Productivity loss}
\format{A data table with 2 variables.
\describe{
  \item{est}{Estimate of the impact of 1-unit change in HAQ on annual lost earnings.}
  \item{se}{Standard error of point estimate of lost earnings.}
}}
\source{
{Wolfe, Frederick, et al. "Household income and earnings losses among 6,396 persons
 with rheumatoid arthritis." The Journal of Rheumatology 32.10 (2005): 1875-1883.}
}
\usage{
prod.loss
}
\description{
Annual earnings loss from a 1-unit change in the HAQ score.
}
\keyword{datasets}
