% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_acr_oprobit}
\alias{sample_acr_oprobit}
\title{Sample ACR response from ordered probit NMA}
\usage{
sample_acr_oprobit(nsims, m, vcov, basedif_mean, basedif_sd, zdif_mean, zdif_sd)
}
\value{
List containing posterior sample of ACR response for each therapy
}
\description{
Sample ACR response from an NMA using a multinomial likelihood and a probit link.
See for example section 3.6 and example 6 in "Nice DSU Technical Support Document 2:
A Generalized Linear Modeling Framework for Pairwise and Network Meta-Analysis of
Randomized Controlled Trials"
}
