% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_betas}
\alias{sample_betas}
\title{Sample from beta distributions}
\usage{
sample_betas(n, shape1, shape2, col_names = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{shape1}{Alpha parameter in beta distribution.}

\item{shape2}{Beta parameter in beta distribution.}

\item{col_names}{Column names to returned matrix.}
}
\value{
Matrix with each column a sample from a beta distribution.
}
\description{
Sample from multiple independent beta distributions. Produces a unique sample of size \code{n} 
using each element in \code{shape1} and \code{shape2}.
}
