% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_dirichlet}
\alias{sample_dirichlet}
\title{Sample matrices using a Dirichlet distribution}
\usage{
sample_dirichlet(n, mat)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{mat}{A matrix where each row is a separte vector of shape parameters.}
}
\value{
An array of matrices representing a sample from the dirichlet distribution for each row.
}
\description{
A wrapper using the function \code{hesim::rdiricihlet_mat}. Maintains the row and column names of \code{mat}.
}
