% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Sample from gamma distributions}
\usage{
sample_gamma(n, mean, se, col_names = NULL)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{mean}{Vector of means.}

\item{se}{Vector of standard errors.}
}
\value{
Matrix with each column a sample from a gamma distribution.
}
\description{
Sample from multiple independent gamma distributions. The parameters of the gamma distribution are derived using methods of moments from
the means and standard errors. Produces a unique sample of size \code{n} using each element in 
\code{mean} and \code{sd}.
}
