% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_gammas}
\alias{sample_gammas}
\title{Sample from gamma distributions}
\usage{
sample_gammas(n, mean = NULL, se = NULL, shape = NULL, rate = NULL,
  col_names = NULL)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{mean}{Vector of means.}

\item{se}{Vector of standard errors.}

\item{col_names}{Column names to returned matrix.}
}
\value{
Matrix with each column a sample from a gamma distribution.
}
\description{
Sample from multiple independent gamma distributions. The parameters of the gamma distribution are derived using methods of moments from
the means and standard errors. Produces a unique sample of size \code{n} using each element in 
\code{mean} and \code{sd}.
}
