% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_mvnorm}
\alias{sample_mvnorm}
\title{Sample from a multivariate normal distribution}
\usage{
sample_mvnorm(n, mu, Sigma)
}
\arguments{
\item{n}{The number of samples required.}

\item{mu}{A vector giving the means of the variables.}

\item{Sigma}{A positive-definite symmetric matrix specifying the covariance matrix of the variables.}
}
\value{
Matrix of samples from the multivariate distribution.
}
\description{
A wrapper for \code{MASS::mvrnorm} that returns a matrix (rather than a vector) if n = 1.
}
