% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_nma_lm}
\alias{sample_nma_lm}
\title{Sample from Bayesian linear model}
\usage{
sample_nma_lm(nsims, m, vcov, k_lower = 1, k_upper = 1, ncovs = 1,
  tx_names = NULL)
}
\arguments{
\item{nsims}{Number of observations.}

\item{m}{Mean for each coefficient.}

\item{vcov}{Variance-covariance matrix of coefficients.}

\item{k_lower}{Lower bound for constant \eqn{k}.}

\item{k_upper}{Upper bound for constant \eqn{k}.}
}
\value{
List containing posterior samples of changes in outcomes.
}
\description{
Sample change in outcomes from Bayesian linear model for NMA.
}
