% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_pars_haq_lcgm}
\alias{sample_pars_haq_lcgm}
\title{Sample parameters for the LCGM}
\usage{
sample_pars_haq_lcgm(nsims, pars)
}
\value{
List containing posterior sample of paramaters from the Norton(2014)
LCGM.
}
\description{
Sample parameters for the latent class growth model used to simulate the progression
of HAQ over time developed by Norton (2014).
}
