% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{sample_pats}
\alias{sample_pats}
\title{Sample patients}
\usage{
sample_pats(n = 1, type = c("homog", "heterog"), age_mean = 55,
  age_sd = 13, male_prop = 0.21, bhaq0_mean = 1.5, bhaq0_sd = 0.7,
  wtmale = 89, wtfemale = 75, dis_dur_mean = 18.65, dis_dur_sd = 12.25,
  prev_dmards_mean = 3.28, prev_dmards_sd = 1.72)
}
\arguments{
\item{n}{Number of samples.}

\item{type}{Should male and female patients be heterogeneous or homogeneous. Default is homogeneous.}

\item{age_mean}{Mean age.}

\item{age_sd}{Standard deviation of age.}

\item{male_prop}{Proportion male.}

\item{bhaq0_mean}{Mean baseline HAQ score.}

\item{bhaq0_sd}{Standard deviation of baseline HAQ score.}

\item{wtmale}{Male weight.}

\item{wtfemale}{Female weight.}

\item{dis_dur_mean}{Mean disease duration}

\item{dis_dur_sd}{Standard deviation of disease duration}

\item{prev_dmards_mean}{Mean number of previous DMARDs}

\item{prev_dmards_sd}{Standard deviation of number of previous DMARDs}
}
\value{
Matrix of patient characteristics. One row for each patient and one column
for each variable. Current variables are:
\describe{
  \item{age}{age in years}
  \item{male}{1 = male, 0 = female}
  \item{haq0}{Baseline HAQ score}
  \item{weight}{Patient weight}
}
}
\description{
Sample patients for rheumatoid arthritis individual patient simulation.
}
