% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_stratified_survpars}
\alias{sample_stratified_survpars}
\title{Sample stratified survival parameters}
\usage{
sample_stratified_survpars(nsims, x)
}
\arguments{
\item{nsims}{Size of the posterior sample.}

\item{x}{A list of survival parameters. For example, see 'Treatment duration' in \link{sample_pars}.}
}
\value{
A list of lists. The top level list is the value of the variable used for 
stratification. The second level list consists of matrices each containing random draws of the
 parameters of the survival distribution. The matrices are for the location parameter and
 each of the ancillary parameters.
}
\description{
Sample survival paramters stratified by variable.
}
\keyword{internal}
