% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_survpars}
\alias{sample_survpars}
\title{Sample treatment duration parameters}
\usage{
sample_survpars(nsims, x)
}
\arguments{
\item{nsims}{Size of the posterior sample.}

\item{x}{A list of treatment duration parameters. See 'Treatment duration' in \link{sample_pars}.}
}
\value{
A list of matrices containing random draws of the parameters of the survival distribution. One matrix
for the location parameter and each of the ancillary parameters.
}
\description{
Generate a random sample of parameters for each probability distribution used to model treatment duration. Parameters are
sampled using a multivariate normal distribution.
}
