% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-structures.R
\name{select_model_structures}
\alias{select_model_structures}
\title{Select model structures}
\usage{
select_model_structures(tx_ihaq = "acr-haq", tx_iswitch = "acr-switch",
  cdmards_haq_model = "lcgm", ttd_cause = "all", ttd_dist = "exponential",
  utility_model = "mixture")
}
\arguments{
\item{tx_ihaq}{Model structure relating treatment to HAQ during the first 6 months of 
treatment. Options, which are equivalent to H1-H3 in the documentation are:
\itemize{
\item{\code{acr-haq}}{ H1: Treatment -> ACR -> HAQ }
\item{\code{acr-eular-haq}}{ H2: Treatment -> ACR -> EULAR -> HAQ}
\item{\code{haq}}{ H3:Treatment -> HAQ}
}}

\item{tx_iswitch}{Model structure relating treatment to switching during the first
6 months of treatment. Options, which are equivalent to S1-S6 in the documentation are:
\itemize{
\item{\code{acr-switch}}{ S1: Treatment -> ACR -> Switch}
\item{\code{acr-das28-switch}}{ S2: Treatment -> ACR -> DAS28 -> Switch}
\item{\code{acr-sdai-switch}}{ S3: Treatment -> ACR -> SDAI -> Switch}
\item{\code{acr-cdai-switch}}{ S4: Treatment -> ACR -> CDAI -> Switch}
\item{\code{das28-switch}}{ S5: Treatment -> DAS28 -> Switch}
 \item{\code{acr-eular-switch}}{ S6: Treatment -> ACR -> EULAR -> Switch}

}}

\item{cdmards_haq_model}{Model used for long-term HAQ progression. Options are:
\itemize{
\item{lcgm}{ Latent class growth model}
\item{linear}{ Constant linear HAQ progression}
}
If \code{lgcm} is chosen, then a latent class growth model is used for cDMARDs
and NBT but a constant annual rate is is assumed for all other therapies; otherwise 
a constant linear HAQ progression is assumed for all therapies including cDMARDs and NBT.}

\item{ttd_cause}{Cause of treatment discontinuation. Options are:
\itemize{
\item{all}{ Treatment discontinuation due to any cause.}
\item{si}{ Treatment discontinuation due to serious infections}
}}

\item{ttd_dist}{Distribution used to model time to treatment discontinuaton. Options are:
\itemize{
\item{exponential}{ Exponential}
\item{weibull}{ Weibull}
\item{gompertz}{ Gompertz}
\item{gamma}{ Gamma}
\item{llogis}{ Log-logistic}
\item{lnorm}{ Lognormal}
\item{gengamma}{ Generalized gamma}
}}

\item{utility_model}{Model used to estimate patient utility as a function of HAQ and patient
characteristics. Options are:
\itemize{
\item{mixture}{ Hernandez Alava (2013) mixutre model}
\item{wailoo}{ Wailoo 2006 logistc regression}
}}
}
\description{
Select the model structures to use in the IVI-RA individual patient simulation.
}
\examples{
mod.structs <- select_model_structures(tx_ihaq = c("acr-haq", "acr-eular-haq"),
                                      tx_iswitch = c("acr-switch", "acr-eular-switch"),
                                      cdmards_haq_model = c("lcgm", "linear"),
                                      ttd_cause = c("all", "si"),
                                      ttd_dist = c("gengamma", "exponential"),
                                      utility_model = c("mixture", "wailoo"))
print(mod.structs)                                       

}
