% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_dhaq6}
\alias{sim_dhaq6}
\title{Simulate change in HAQ at 6 months}
\usage{
sim_dhaq6(treatments, input_data, pars, tx_lookup = iviRA::treatments$sname,
  hist = c("naive", "experienced"), line = 1, tx_ihaq = c("acr-haq",
  "acr-eular-haq", "haq"))
}
\arguments{
\item{treatments}{Vector of treatments arms to simulate. Should correpond to \code{sname} in 
\link{iviRA::treatments}.}

\item{input_data}{An object of class \code{input_data} returned from \link{get_input_data}. 
The only elements required are \code{x_acr} and \code{x_haq}.}

\item{pars}{List of sampled parameter values generated from \link{get_input_data}. The only 
elements required are \code{acr}, \code{haq}, \code{acr2haq}, \code{acr2eular}, and
\code{eular2haq}.}

\item{tx_lookup}{Vector of treatments with names equivalent to \code{iviRA::treatments$sname}.
Index of treatments in \code{treatments} are matched against treatments in \code{tx_lookup} by 
name. Indices of parameter estimates from the network meta-analyses bust be in the same order as
in \code{tx_data}.}

\item{hist}{Is the patient biologic naive or biologic experienced?}

\item{line}{Line of therapy}

\item{tx_ihaq}{Equivalent to argument \code{tx_haq} in \link{select_model_structures}.}
}
\value{
Change in HAQ for each selected model structure, randomly sampled parameter set,
simulated patient, and treatment.
}
\description{
Simulate change in HAQ at 6 months using different model structures relating treatment to HAQ.
}
