% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_haq}
\alias{sim_haq}
\title{Simulate HAQ over time}
\usage{
sim_haq(arminds, input_data, pars, dur_dist = "lnorm", si_dist = "exp",
  max_months = NULL, nbt_ind = which(therapy.pars$info$sname == "nbt"),
  check = TRUE)
}
\arguments{
\item{arminds}{Indices of treatment arms consisting of sequences of therapies (each element
is the index of a therapy in \code{therapy.pars$info}. May be a vector consisting of a single treatment sequence or a matrix 
of unique sequences for each patient.}

\item{input_data}{List of input data. Required inputs are \code{haq0}, \code{age}, \code{male}, \code{x.mort}, 
and \code{x.dur} as generated from \link{input_data}.}

\item{pars}{List of parameters. Required parameters are \code{rebound}, \code{acr1}, \code{acr2}, \code{acr2eular}, \code{eular2haq}, 
\code{haq.lprog.therapy}, \code{haq.lprog.age}, \code{logor.mort}, \code{mort.loghr.haqdif}, \code{si.surv},
\code{dur.eular.mod}, \code{dur.eular.good
}, and \code{lt} as
generated from \link{sample_pars}.}

\item{dur_dist}{Survival distribution for treatment duration.}

\item{si_dist}{Survival distribution for serious infections. Currently rate is assumed
to be constant so an exponential distribution is used.}

\item{max_months}{Maximum number of months to run the model for. Default is NULL which implies that
the model is simulated over each patient's lifetime.}

\item{nbt_ind}{Index for the non-biologic (i.e. a term used to define a selection of treatments clinicians use
after the last biologic in a treatment sequence).}

\item{check}{Should the function check parameters and input data passed to the model? Default is TRUE.}
}
\value{
A \code{data.table} with the following columns:
\describe{
\item{sim}{Simulation number. Indexed from 1 to S where S is the number of randomly sampled parameter sets (e.g. n from \link{sample_pars}).}
\item{id}{ID number. Indexed from 1 to N where N is the number of simulated patients (e.g. n from \link{sample_pats}).}
\item{month}{Month since a simulated patient began the first therapy in a treatment sequence.}
\item{therapy}{Index of therapy used. Given J total therapies, the first J - 1 therapies match the indices from \code{arminds}. The final \code{therapy}
 is always equal to the index of the non-biologic therapy (\code{nbt_ind}).}
\item{therapy_seq}{Number of therapies in a sequence. First therapy equal to 1, second therapy equal to 2, \ldots}
\item{therapy_cycle}{Number of model cycles since a patient began taking a given therapy in a treatment sequence. \code{therapy_cycle} = 1
during the initial 6-month treatment period.}
\item{death}{Equal to 1 if patient died during the model cycle and 0 otherwise. }
\item{age}{Age of patient which increases with the model cycles.}
\item{ttd}{Time to treatment discontinuation. Measured in terms of model cycles (e.g. ttd = 2 if therapy will discontinue in 
1 year given 6-months cycles). \code{ttd} is measured at the end of each cycle. Patients switch therapies during the cycle in which \code{ttd} 
becomes negative and HAQ rebounds during the cycle.}
\item{acr}{Simulated ACR response during the initial 6-month period for a new therapy. Constant within \code{therapy}. 
Categories are 0 (ACR < 20), 1 (ACR 20-50), 2 (ACR 50-70), and 3 (ACR 70+).}
\item{eular}{Simulated EULAR response during the initial 6-month period for a new therapy. Constant within \code{therapy}. 
Categories are 0 (no EULAR response), 1 (moderate EULAR response), and 2 (good EULAR response).}
\item{haq}{HAQ score. Restricted to range between 0 and 3.}
\item{ttsi}{Time to serious infection. Like \code{ttd}, measured in terms of model cycles. \code{ttsi} is measured at the end of each 
cycle.}
\item{si}{Equal to 1 if treatment discontinuation was caused by a serious infection and 0 otherwise.}
\item{yrlen}{Length of a model cycle in years. Equal to 0.5 given 6-month cycles.}
}
}
\description{
An individual patient simulation of HAQ scores for rheumatoid arthritis patients given a sequence of J 
treatments. An "outer" loop iterates over S draws from sampled parameter sets and an "inner" loop iterates of
N patients. Cycle lengths are 6 months long. The simulation is written in C++ for speed.
}
