% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_prod_loss}
\alias{sim_prod_loss}
\title{Simulate productivity loss from simulated HAQ score}
\usage{
sim_prod_loss(simhaq, pl_haq, check = TRUE)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}.}

\item{pl_haq}{Decrease in wages (i.e. productivity loss) per one-unit increase in HAQ. Equivalent to output \code{prod.loss} from
\link{sample_pars}.}
}
\value{
Vector of simulated productivity loss for each simulated patient and time-period.
}
\description{
Simulate productivity loss from HAQ score simulated using \code{sim_haq}
}
