% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_utility}
\alias{sim_utility}
\alias{sim_utility_mixture}
\alias{sim_utility_wailoo}
\alias{sim_qalys}
\title{Simulate utility after IPS}
\usage{
sim_utility_mixture(simhaq, male, pars)

sim_utility_wailoo(simhaq, haq0, male, prev_dmards, coefs)

sim_qalys(simhaq, utility, si_ul, x_attr, tx_attr_coef)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}. Must include columns \code{yrlen} for
year length of model cycle, \code{sim} for simulation number, and \code{si} for whether a serious
infection occured during the model cycle.}

\item{male}{Indicator = 1 for males and 0 for females.}

\item{pars}{List of sampled parameters needed to simulate utility using the Hernandez Alava (2013) mixture 
model (i.e., the element \code{utility.mixture} returned by \link{sample_pars}).
to \link{sample_pars}.}

\item{haq0}{HAQ score at baseline.}

\item{prev_dmards}{Number of previous DMARDs.}

\item{coefs}{Matrix of sampled coefficients needed to simulate utility using the Wailoo (2006) model 
(i.e., the element \code{utility.wailoo} returned by \link{sample_pars}). Note that the matrix 
columns must contain the exact same variables as generated by \link{sample_pars}.}

\item{utility}{Simulated utility from \code{sim_utility_mixture} or \code{sim_utility_wailoo}.}

\item{si_ul}{Sampled utility loss. Equivalent to output \code{si.ul} from \link{sample_pars}.}

\item{x_attr}{Treatment attribute data (e.g., ouptut \code{x.attr} from
\link{get_input_data})}

\item{tx_attr_coef}{Distribution of coefficient vector for treatment attributes (e.g., output
\code{utility.tx.attr} from \link{sample_pars}.)}

\item{simhaq}{Simulation output from \link{sim_iviRA}. Variables needed are \code{sim}, \code{id},
\code{age}, and \code{haq}.}
}
\value{
For \code{sim_utility_mixture} and \code{sim_utility_wailoo}, a vector of 
simulated utility for each row returned in \code{simhaq}. For \code{sim_qalys}, a vector
of QALYs for each row in \code{simhaq}.
}
\description{
Simulate utility after running \link{sim_iviRA} with \code{output = "data"}. This can be useful 
in cases where you want to use a different algorithm to estimate utility, but do not want to rerun 
the entire simulation.
}
\details{
Note that disease duration is set to 18.65 years in \code{sim_utility_wailoo}, which
is the mean value from the Wailoo (2006) paper used for the parameter estimates. Age and 
the HAQ score are taken from the simulation output.
}
\examples{
pop <- sample_pop(n = 10)
tx.seq <- c("adamtx", "cdmards")
mod.structs <- select_model_structures(utility_model = "wailoo")
input.dat <- get_input_data(pop = pop)
parsamp <- sample_pars(n = 10, input_dat = input.dat)
sim.out <- sim_iviRA(tx_seqs = tx.seq, input_data = input.dat, pars = parsamp,
                    model_structures = mod.structs, output = "data")
utility.mix <- sim_utility_mixture(simhaq = sim.out, male = pop[, "male"], pars = parsamp$utility.mixture)
utility.wailoo <- sim_utility_wailoo(simhaq = sim.out, haq0 = pop[, "haq0"], male = pop[, "male"],
                                prev_dmards = pop[, "prev_dmards"], 
                                coefs = parsamp$utility.wailoo)
qalys.mix <- sim_qalys(simhaq = sim.out, utility = utility.mix, si_ul = parsamp$si.ul,
                       x_attr = input.dat$x.attr, tx_attr_coef = parsamp$utility.tx.attr)
head(utility.mix)
head(utility.wailoo)   
head(qalys.mix)

}
