% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_utility_mixture}
\alias{sim_utility_mixture}
\title{Simulate utility from simulated HAQ score (Hernandez-Alva)}
\usage{
sim_utility_mixture(simhaq, male, pars, check = TRUE)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}.}

\item{male}{Vector indiciating patient gender (1 = male, 0 = female).}

\item{pars}{List of parameters needed to simulate utility using the Hernandez (2013) mixture 
model. These are the parameters 'mixture.utility' described in the documentation 
to \link{sample_pars}.}

\item{check}{Should the function check parameters and input data passed to the model? Default is TRUE.}
}
\value{
Matrix. First column is simulated pain score and second column is simulated utility. 
Each row corresponds to a unique patient and time-period (i.e. month) from \link{sim_haq}.
}
\description{
Simulate utility from HAQ score simulated using \code{sim_haq} using mixture model from
Hernandez-Alva (2013).
}
