% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_utility_wailoo}
\alias{sim_utility_wailoo}
\title{Simulate utility from simulated HAQ score using Wailoo (2006) model}
\usage{
sim_utility_wailoo(simhaq, input_data, pars, check = TRUE)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}. Variables needed are \code{sim}, \code{id},
\code{age}, and \code{haq}.}

\item{input_data}{Variables other than those in simhaq needed to simulate model; these are
\code{Baseline HAQ}, \code{Male}, and \code{Previous DMARDs}. All variables
are generated from \link{input_data}. Note that disease duration is set to 18.65 years, which
is the mean value from the Wailoo (2006) paper used for the parameter estimates.}

\item{pars}{Matrix of coefficients needed to simulate utility using the Wailoo (2006) model. 
See the documentation in \link{sample_pars} for details. Note that the matrix columns must be in 
the same order as generated by \link{sample_pars}.}

\item{check}{Should the function check parameters and input data passed to the model? Default is TRUE.}
}
\value{
Vector of utility scores.
}
\description{
Simulate utility from HAQ score simulated using \code{sim_haq} using model from
Wailoo (2006).
}
