% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{survival_summary}
\alias{survival_summary}
\title{Survival model parameters}
\usage{
survival_summary(x)
}
\value{
matrix
}
\description{
Summarize parameters from a survival model
}
\keyword{internal}
