% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treat.cost}
\alias{treat.cost}
\title{Treatment costs}
\format{A data table with 5 variables.
\describe{
  \item{name}{Long-form name of treatment.}
  \item{sname}{Short-form name of treatment.}
  \item{dosage}{FDA approved dosage.}
  \item{strength_dosage_form}{Strength and dosage form of treatment}
  \item{init_dose_val}{Numeric value of dose approved by FDA during the first 6 months of
  treatment.}
  \item{ann_dose_val}{Numeric value of dose approved by FDA beyond the first 6 months.}
  \item{dose_unit}{Unit of FDA approved dose.}
  \item{init_num_doses}{Number of doses based on FDA used during the first 6 months.}
  \item{ann_num_doses}{Annual number of doses based on FDA beyond the first 6 months.}
  \item{strength_val}{Numeric value of strength of treatment.}
  \item{strength_unit}{Unit of treatment.}
  \item{wac_per_unit}{Wholesale acquisition cost per unit of treatment.}
  \item{infusion_cost}{Cost per infusion.}
}}
\usage{
treat.cost
}
\description{
Drug acquisition and administration costs for medications used to treat rheumatoid
arthritis.
}
\keyword{datasets}
