% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treat.switch}
\alias{treat.switch}
\title{Treatment switching at 6 months}
\format{A data.table containing the following columns:
\describe{
  \item{variable}{Name of variable. Includes the intercept, an indicator variable for whether
  disease activity is moderate (CDAI 10-22), and an indicator variable for 
  whether disease activity is high (CDAI greater than 22).}
  \item{or}{The odds ratio of the coefficients.}
  \item{or_lower}{Lower 95 percent confidence interval for the odds ratio.}
  \item{or_upper}{Upper 95 percent confidence interval for the odds ratio.}
  \item{logor}{Log odds ratio of the coefficents.}
  \item{logor_lower}{Lower 95 percent confidence interval for the log of the odds ratio.}
  \item{logor_upper}{Upper 95 percent confidence interval for the log of the odds ratio.}
}}
\source{
Zhang, Jie, et al. "Thresholds in disease activity for switching biologics in rheumatoid 
arthritis patients: experience from a large US cohort." 
Arthritis care & research 63.12 (2011): 1672-1679.
}
\usage{
treat.switch
}
\description{
Regression coefficients from a logistic regression equation predicting the probability of 
switching treatment at 6 months by disease activity level.
}
\keyword{datasets}
