% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttd}
\alias{ttd}
\alias{ttd.eular.mod}
\alias{ttd.eular.good}
\alias{ttd.eular.mod.adj}
\alias{ttd.eular.good.adj}
\title{Time to treatment discontinuation parameters}
\format{A list of lists. One list for each parametric distribution. Within each distribution
the list contains parameters produced by \code{flexsurv} on real-line scale
 (e.g. after log transformation if they are defined as positive). The parameters are:
\describe{
  \item{coef}{Coefficients for all parameters including location parameter and 
  ancillary parameter(s)}
  \item{vcov}{Covariance matrix for all parameters including location parameter and 
  ancillary parameter(s)}
}}
\source{
Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, 
golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously
 treated with disease-modifying antirheumatic drugs and after the failure of conventional 
 disease-modifying antirheumatic drugs only: systematic review and economic evaluation." 
Health Technology Assessment 20.35 (2016): 1-610.
}
\usage{
ttd.eular.mod

ttd.eular.good

ttd.eular.mod.adj

ttd.eular.good.adj
}
\description{
A list of parameters estimated in a parametric survival analysis of time to treatment 
discontinuation. Models were fit using \emph{flexsurv}. Survival curves were scanned from the 2016 NICE  
HTA and then converted to individual-level patient dat using the \code{reconstruct_ipd}
function. Survival curves are stratified by patients with moderate and good responses 
to treatment according to Eular response categories.
}
\keyword{datasets}
