% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttsi}
\alias{ttsi}
\title{Time to serious infection parameters}
\format{A data.table with three columns:
\describe{
\item{sname}{Short name of treatment.}
\item{lograte}{Log of serious infection rate.}
\item{lograte_se}{Standard error of log of serious infection rate.}
}}
\source{
Singh, Jasvinder A., et al. "Adverse effects of biologics: a network meta-analysis 
and Cochrane overview." Cochrane Database of Systematic Reviews 2 (2010).
}
\usage{
ttsi
}
\description{
Parameters for sampling serious infections using an exponential distribution.
}
\keyword{datasets}
