% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx.attr}
\alias{tx.attr}
\title{Treatment attributes}
\format{A list with two elements. 
\describe{
  \item{data}{A matrix where each column is a treatment attribute and each row corresponds
  to a treatment.}
  \item{utility.gain}{A matrix with two columns, lower and upper, corresponing to the lower
  and upper bounds of the utility gain associated with each treatment attribute in the 'data' 
  matrix.}
}}
\source{
None. Currently determined by the user.
}
\usage{
tx.attr
}
\description{
Impact of treatment attributes other than safety and efficacy on utility.
}
\keyword{datasets}
