% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{util.wailoo.pars}
\alias{util.wailoo.pars}
\title{Wailoo (2006) utility mapping}
\format{List containing vectors of coefficients and standard errors}
\source{
Wailoo, A., et al. "Modeling the cost effectiveness of etanercept, adalimumab and 
anakinra compared to infliximab in the treatment of patients with rheumatoid arthritis in the
 Medicare program." Rockville, MD: Agency for Healthcare Research and Quality (2006).
}
\usage{
util.wailoo.pars
}
\description{
Coefficients and standard errors from Wailoo (2006) model
used to map HAQ score to utilities.
}
\keyword{datasets}
