% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utility.mixture}
\alias{utility.mixture}
\title{Mixture model utility mapping}
\format{List containing coefficient vector (point estimates and parameter names) and 
variance-covariance matrix, and indices for parameters.}
\source{
Alava, Monica Hernandez, et al. "The relationship between EQ-5D, 
 HAQ and pain in patients with rheumatoid arthritis." Rheumatology 52.5 (2013): 944-950.
}
\usage{
utility.mixture
}
\description{
Coefficients and variance-covariance matrix from the Alava (2013) mixture model
used to map HAQ score to utilities.
}
\keyword{datasets}
