% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utility.wailoo.pars}
\alias{utility.wailoo.pars}
\title{Wailoo (2006) utility mapping}
\format{A data.table containing the following columns.
\describe{
  \item{var}{Name of the variable.}
  \item{est}{Point estimate on the log odds scale.}
  \item{loghr_se}{Standard error of the point estimate.}
}}
\source{
Wailoo, A., et al. "Modeling the cost effectiveness of etanercept, adalimumab and 
anakinra compared to infliximab in the treatment of patients with rheumatoid arthritis in the
 Medicare program." Rockville, MD: Agency for Healthcare Research and Quality (2006).
}
\usage{
utility.wailoo.pars
}
\description{
Coefficients and standard errors from Wailoo (2006) logistic regression equation
 used to map HAQ score to utilities.
}
\keyword{datasets}
