% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acr2eular}
\alias{acr2eular}
\title{ACR to Eular Conversion}
\format{A matrix where each row shows the probability that a particualar responder in an 
ACR category will have a given Eular response. Probabilities are based on observed 
relationships in the VARA database.}
\source{
Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, 
golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously
 treated with disease-modifying antirheumatic drugs and after the failure of conventional 
 disease-modifying antirheumatic drugs only: systematic review and economic evaluation." 
Health Technology Assessment 20.35 (2016): 1-610.
}
\usage{
acr2eular
}
\description{
Convert ACR response (ACR < 20, ACR20, ACR50, ACR70) to Eular response (none, moderate, good).
}
\keyword{datasets}
