% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acr2haq}
\alias{acr2haq}
\title{HAQ change by 6-month ACR Response}
\format{A data.table with rows as ACR response category. First column is mean and second 
column is standard deviation.}
\source{
Carlson, Josh J., et al. "Economic evaluation of tocilizumab monotherapy compared to adalimumab monotherapy 
in the treatment of severe active rheumatoid arthritis." Value in Health 18.2 (2015): 173-179.
}
\usage{
acr2haq
}
\description{
Mean and standard deviation of HAQ change by 6-month ACR response category.
}
\keyword{datasets}
