% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{check_sim_haq}
\alias{check_sim_haq}
\title{Check parameters of sim_haq}
\usage{
check_sim_haq(input_data, pars)
}
\arguments{
\item{input_data}{\code{input_data} as passed to \link{sim_haq}.}

\item{pars}{\code{pars} as passed to \link{sim_haq}.}
}
\description{
Error messages when incorrect inputs are passed to sim_haq.
}
\keyword{internal}
