% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{get_input_data}
\alias{get_input_data}
\title{Input data for simulation}
\usage{
get_input_data(pop, x_acr = NULL, x_haq = NULL, x_das28 = NULL,
  x_ttd_all = NULL, x_ttd_da = NULL, x_ttd_eular = NULL, x_mort = NULL,
  x_attr = iviRA::utility.tx.attr$x)
}
\arguments{
\item{pop}{The patient population. A matrix that must contain variables generated from 
\link{sample_pop}: age' for age, 'haq0' for baseline HAQ, 'male' as a indicator equal to
1 if the patient is male and 0 if female, 'weight' for patient weight, 'prev_dmards' 
for number of previous DMARDs, 'das28' for the patient's DAS28 score, 'sdai' for the 
patient's SDAI score, and 'cdai' for the patient's CDAI score.}

\item{x_acr}{Design matrix where each column is a variable known at baseline that is 
used to predict the relative treatment effects for ACR response. By default, 
only includes an intercept, which implies  that there are no treatment-by-covariate 
interactions.}

\item{x_haq}{Design matrix where each column is a variable known at baseline that is
used to predict the relative treatment effects for change in HAQ at 6 months from baseline. 
By default, only includes an intercept, which implies that there are no 
treatment-by-covariate interactions.}

\item{x_das28}{Design matrix where each column is a variables known at baseline that 
is used to predict the relative treatment effects for change in DAS28 at 6 months from 
baseline. By default, only includes an intercept, which implies that there are no
 treatment-by-covariate interactions.}

\item{x_ttd_all}{Design matrix where each column is a variable influencing treatment duration
in a model representative of all patients. The impact of each variable is determined
by the sampled values of the coefficients used to predict the location parameter in 
\code{ttd.all} returned by \link{sample_pars}.}

\item{x_ttd_da}{Design matrix where the first column is the intercept, the second column
is a dummy variable used to indicate whether a patient has moderate disease activity, and
the third column is dummy variable used to indicate whether a patient has high disease 
activity (note, however, that the second and third columns are updated during the simulation
according to the simulated disease activity level). All remaining columns after the 3rd column
are variables influencing treatment duration. The impact of each variable is determined 
by the sampled values of the coefficients used to predict the location parameter in 
 \code{ttd.da} returned by \link{sample_pars}.}

\item{x_ttd_eular}{Design matrix where each column is a variable influencing
treatment duration in models stratified by EULAR response. The impact of each variable
is determined by the sampled values of the coefficients used to predict the location
 parameter in \code{ttd.eular} returned by \link{sample_pars}.}

\item{x_mort}{Design matrix where each column is a variable influencing mortality. The impact 
of each variable is determined by the parameter vector \code{mort.logor} returned by
\link{sample_pars}.}

\item{x_attr}{Design matrix where each column is a variable related to treatment attributes
related to the processes of care influencing utility. The impact 
of each variable is determined by the parameter vector \code{tx.attr.utility} returned by
\link{sample_pars}}
}
\value{
A list containing the following data inputs:
\describe{
  \item{n}{Number of simulated patients.}
  \item{haq0}{A vector of patient HAQ at baseline.}
  \item{age}{A vector of patient age at baseline.}
  \item{male}{A vector of patient gender (1 = male, 0 = female).}
  \item{prev.dmards}{A vector of the number of previous DMARDs.}
  \item{x.acr}{Equivalent to \code{x.acr} passed as an argument to the function.}
  \item{x.haq}{Equivalent to \code{x.haq} passed as an argument to the function.}
  \item{x.das28}{Equivalent to \code{x.das28} passed as an argument to the function.}
  \item{x.ttd.all}{Equivalent to \code{x.ttd.all} passed as an argument to the function.}
  \item{x.ttd.da}{Equivalent to \code{x.ttd.da} passed as an argument to the function.}
 \item{x.ttd.eular}{Equivalent to \code{x.ttd.eular} passed as an argument to the function.}
  \item{x.mort}{Equivalent to \code{x.mort} passed as an argument to the function.}
  \item{x.attr}{Equivalent to \code{x.attr} passed as an argument to the function.}
}
}
\description{
Generate data inputs for the the individual patient simulation (\link{sim_iviRA}).
}
\details{
If a design matrix is set to NULL, then a single column of ones is returned. 
In other words, if a design matrix is not specified, then it is assumed that an intercept
 only model will be used.
}
\examples{
pop <- sample_pop(n = 100)
input.dat <- get_input_data(pop)
names(input.dat)
head(input.dat$haq0)
head(input.dat$x.haq)
}
