% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence}
\alias{incidence}
\alias{incidence.female}
\alias{incidence.male}
\title{Incidence}
\format{A data frame with 86 rows and 4 variables:
\describe{
  \item{age}{Age in years.}
  \item{events}{Number of events occuring between ages x and x + 1.}
  \item{person_years}{Number of person years between ages x and x + 1.}
  \item{incidence_rate}{Incidence rate between ages x and x + 1.}

}}
\source{
Myasoedova, Elena, et al. "Is the incidence of rheumatoid arthritis rising? Results from
Olmsted Country, Minnesota, 1995-2007." Arthritis & Rheumatology 62.6 (2010): 1576-1582.
}
\usage{
incidence.female

incidence.male
}
\description{
Incidence of rheumatoid arthritis in the United States.
}
\details{
The incidence rate for each age is derived from Table 2 in Myasoedova et al. (2010), which
was a study conducted in Olmsted County, Minnesota for residents age 18 years of age and older
between 1995 and 2007.
}
\keyword{datasets}
