% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{input_data}
\alias{input_data}
\title{Input data for simulation}
\usage{
input_data(patdata, vars_mort = "haq0", vars_dur = "1")
}
\arguments{
\item{patdata}{Matrix of patient data. Must contain variables generated from \link{sample_pats}:
 'age' for age, 'haq0' for baseline HAQ, 'male' as a indicator equal to
1 if the patient is male and 0 if female, 'weight' for patient weight, 'dis_dur' for disease
duration and 'prev_dmards' for number of previous DMARDs.}

\item{vars_mort}{A chararacter vector of variables in \code{patdata} to be used to adjust
'qx' (using odds ratios) in the lifetables. Use '1' to include an intercept.}

\item{vars_dur}{A character vector of variables in \code{patdata} to be used to predict treatment
duration. Use '1' to include an intercept.}
}
\value{
A list containing the following data inputs:
\describe{
  \item{n}{Number of simulated patients}
  \item{haq0}{A vector of patient HAQ at baseline.}
  \item{age}{A vector of patient age at baseline.}
  \item{male}{A vector of patient gender (1 = male, 0 = female).}
  \item{dis.dur}{A vector of disease duration}
  \item{prev.dmards}{A vector of the number of previous DMARDs}
  \item{x.mort}{Design matrix for mortality adjustment with odds ratios}
  \item{x.ttd}{Design matrix for treatment duration.}
}
}
\description{
Generate data inputs for the IPS.
}
