% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi-ra.R
\docType{package}
\name{iviRA}
\alias{iviRA}
\alias{iviRA-package}
\title{iviRA: an R package for a health-economic rheumatoid arthritis model}
\description{
iviRA: an R package for a health-economic rheumatoid arthritis model
}
