% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{mom_beta}
\alias{mom_beta}
\title{Method of moments for Beta Distribution}
\usage{
mom_beta(mean, sd, quant = NULL)
}
\value{
parameters of beta distribution
}
\description{
Estimate parameters of Beta distribution from mean
and variance.
}
