% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort.or}
\alias{mort.or}
\title{Mortality odds ratios by patient characteristics}
\format{A data frame with 1 row and 9 columns:
\describe{
  \item{var}{Name of variable}
  \item{or}{Odds ratio without radiographic data}
  \item{or_se}{Standard error of odds ratio without radiographic data}
  \item{logor}{Log odds ratio without radiographic data}
  \item{logor_se}{Standard error of log odds ratio without radiographic data}
}}
\source{
{Wolfe, Frederick, et al. "Predicting mortality in patients with rheumatoid arthritis." 
Arthritis & Rheumatism 48.6 (2003): 1530-1542.}
}
\usage{
mort.or
}
\description{
Impact of HAQ on odds ratio for mortality from table 4 in Wolfe et al (2003).
}
\details{
The standard errors of the log odds ratios are derived from the standard error of the 
odds ratios. By the delta method, \eqn{var(OR) = exp(\beta)var(\beta)exp(\beta)} where 
\eqn{\beta} is the regression coefficient of the logistic regression. Since \eqn{var(\beta)} 
is the variance of the log odds ratio, it follows that
 \eqn{var[log(OR)] = [var(OR)/exp(\beta)}]^2.
}
\keyword{datasets}
