% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-prob.R
\name{ologit_prob}
\alias{ologit_prob}
\title{Bayesian predicted probabilities for ordered logistic regression}
\usage{
ologit_prob(x, beta, cut)
}
\arguments{
\item{x}{Matrix containing variables in the model (i.e. the design matrix)}

\item{beta}{Matrix of coefficients where each row contains random draws of the coefficients
from their posterior distribution}

\item{cut}{Matrix of cutpoints, one for each category, where each row contains random draws
from their posterior distribution}
}
\value{
Matrix of predicted probabilities for each individual and simulation.
}
\description{
Posterior distribution of predicted probabilitities from an ordered logistic regression
}
