% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_nma_acr}
\alias{sample_nma_acr}
\title{Sample ACR response from ordered probit NMA}
\usage{
sample_nma_acr(nsims, m, vcov, k_lower = 1, k_upper = 1, ncovs = 1,
  tx_names = NULL)
}
\arguments{
\item{nsims}{Number of observations.}

\item{m}{Mean for each coefficient.}

\item{vcov}{Variance-covariance matrix of coefficients.}

\item{k_lower}{Lower bound for constant \eqn{k}.}

\item{k_upper}{Upper bound for constant \eqn{k}.}

\item{ncovs}{Number of treatment by covariate interactions.}
}
\value{
List containing posterior sample of ACR response for each therapy
}
\description{
Sample ACR response from an NMA using a multinomial likelihood and a probit link.
See for example section 3.6 and example 6 in "Nice DSU Technical Support Document 2:
A Generalized Linear Modeling Framework for Pairwise and Network Meta-Analysis of
Randomized Controlled Trials"
}
