% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_nma_lm}
\alias{sample_nma_lm}
\title{Sample from Bayesian linear model}
\usage{
sample_nma_lm(nsims, m, vcov, rr_lower, rr_upper, hist, tx_names)
}
\arguments{
\item{nsims}{Number of observations.}

\item{m}{Mean for each coefficient.}

\item{vcov}{Variance-covariance matrix of coefficients.}

\item{rr_lower}{Lower bound for relative risk.}

\item{rr_upper}{Upper bound for relative risk.}

\item{hist}{Patient history equivalent to tx_hist in \link{sample_pars}}
}
\value{
List containing posterior samples of changes in outcomes.
}
\description{
Sample change in outcomes from Bayesian linear model for NMA.
}
