% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_pars}
\alias{sample_pars}
\title{Sample parameters}
\usage{
sample_pars(n = 100, rebound_lower = 0.7, rebound_upper = 1,
  ltfemale = iviRA::lifetable.female, ltmale = iviRA::lifetable.male,
  acr2eular_mat = iviRA::acr2eular, tx_cost = iviRA::tx.cost,
  mort_logor = iviRA::mort.or$logor,
  mort_logor_se = iviRA::mort.or$logor_se,
  mort_loghr_haqdif = iviRA::mort.hr.haqdif$loghr,
  mort_loghr_se_haqdif = iviRA::mort.hr.haqdif$loghr_se,
  ttd_all = iviRA::ttd.all, ttd_da = iviRA::ttd.da,
  ttd_eular = iviRA::ttd.eular, nma_acr_mean = iviRA::nma.acr.naive$mean,
  nma_acr_vcov = iviRA::nma.acr.naive$vcov, nma_acr_rr_lower = 0.75,
  nma_acr_rr_upper = 0.92, nma_das28_mean = iviRA::nma.das28.naive$mean,
  nma_das28_vcov = iviRA::nma.das28.naive$vcov, nma_das28_rr_lower = 0.75,
  nma_das28_rr_upper = 0.92, nma_haq_mean = iviRA::nma.haq.naive$mean,
  nma_haq_vcov = iviRA::nma.haq.naive$vcov, nma_haq_rr_lower = 0.75,
  nma_haq_rr_upper = 0.92, tx_hist = c("naive", "exp"),
  haq_lprog_tx_mean = iviRA::haq.lprog$tx$est,
  haq_lprog_tx_se = iviRA::haq.lprog$tx$se, haq_lcgm_pars = iviRA::haq.lcgm,
  eular2haq_mean = iviRA::eular2haq$mean,
  eular2haq_se = iviRA::eular2haq$se, acr2haq_mean = iviRA::acr2haq$mean,
  acr2haq_se = iviRA::acr2haq$se,
  acr2sdai_lower = iviRA::acr2sdai$inception$lower,
  acr2sdai_upper = iviRA::acr2sdai$inception$upper,
  acr2cdai_lower = iviRA::acr2cdai$inception$lower,
  acr2cdai_upper = iviRA::acr2cdai$inception$upper,
  acr2das28_lower = iviRA::acr2das28$inception$lower,
  acr2das28_upper = iviRA::acr2das28$inception$upper,
  haq_lprog_age_mean = iviRA::haq.lprog$diff.age$est,
  haq_lprog_age_se = iviRA::haq.lprog$diff.age$se,
  hosp_days_mean = iviRA::hosp.cost$days_mean,
  hosp_days_se = iviRA::hosp.cost$days_se,
  hosp_cost_mean = iviRA::hosp.cost$cost_pday_mean,
  hosp_cost_se = iviRA::hosp.cost$cost_pday_se,
  mgmt_cost_mean = iviRA::mgmt.cost$est, mgmt_cost_se = iviRA::mgmt.cost$se,
  pl_mean = iviRA::prod.loss$est, pl_se = iviRA::prod.loss$se,
  ttsi = iviRA::ttsi, si_cost = 5873, si_cost_range = 0.2,
  si_ul = 0.156, si_ul_range = 0.2, tx_names = iviRA::treatments$sname,
  util_mixture_pain = iviRA::pain)
}
\arguments{
\item{n}{Size of the posterior sample.}

\item{rebound_lower}{The rebound is the increase in HAQ following treatment. It is defined as 
a proportion \emph{f} times the size of the inititial treatment response. \code{rebound_lower} defines the
lower bound for \emph{f}. Default is 0.7, which implies
that the rebound post treatment is 0.7 times the initial treatment effect.}

\item{rebound_upper}{\code{rebound_upper} defines the
upper bound for \emph{f}. Default is 1, which implies
that the rebound post treatment is the same as the initial treatment effect.}

\item{ltfemale}{Lifetable for women. Must contain column 'age' for single-year of age and 'qx' for
the probability of death at a given age.}

\item{ltmale}{Identical to \code{ltfemale} but for men.}

\item{acr2eular_mat}{A two-way frequency matrix with columns denoting EULAR response
(none, moderate, good) and rows denoting ACR response  (<20, 20-50, 50-70, 70+).}

\item{tx_cost}{Treatment cost matrix and treatment lookup in format of iviRA::tx.cost.}

\item{mort_logor}{Log odds ratio of impact of baseline HAQ on probability of mortality.}

\item{mort_logor_se}{Standard error of log odds ratio of impact of baseline HAQ on probability of mortality.}

\item{mort_loghr_haqdif}{Log hazard ratio of impact of change in HAQ from baseline on mortality rate. A vector with
each element denoting (in order) hazard ratio for months 0-6, >6 - 12, >12 - 24, >24 -36, >36.}

\item{mort_loghr_se_haqdif}{Standard error of log hazard ratio of impact of change in HAQ from baseline on mortality rate.}

\item{ttd_all}{A list containing treatment duration parameters representative of all patients (i.e., unstratified).
See 'Treatment duration'.}

\item{ttd_da}{A list containing treatment duration parameters. Covariates for moderate and high
disease activity. See 'Treatment duration'.}

\item{ttd_eular}{A list containing treatment duration parameters stratified by EULAR response. See 'Treatment duration'.}

\item{nma_acr_mean}{Posterior means for ACR response NMA parameters on probit scale for 
biologic naive patients (i.e., 1st line). ACR response is modeled using an ordered probit model.}

\item{nma_acr_vcov}{Variance-covariance matrix for ACR response NMA parameters on probit scale for 
biologic naive patients (i.e., 1st line). ACR response is modeled using an ordered probit model.}

\item{nma_acr_rr_lower}{Lower bound for proportion reduction (i.e., relative risk) in 
overlapping ACR response probabilities (ACR20/50/70) for biologic experienced patients 
(i.e., lines 2 and later).}

\item{nma_acr_rr_upper}{Upper bound for proportion reduction (i.e., relative risk) in
overlapping ACR response probabilities (ACR20/50/70) for biologic experienced patients 
(i.e., lines 2 and later).}

\item{nma_das28_mean}{Posterior means for DAS28 NMA parameters for biologic naive 
patients (i.e., 1st line). Change in DAS28 from baseline is modeled using a linear model.}

\item{nma_das28_vcov}{Variance-covariance matrix for DAS28 NMA paramters for biologic naive
patients (i.e., 1st line). Change in DAS28 from baseline is modeled using a linear model.}

\item{nma_das28_rr_lower}{Lower bound for proportion reduction (i.e., relative risk) in DAS28
for biologic experienced patients (i.e., lines 2 and later).}

\item{nma_das28_rr_upper}{Upper bound for proportion reduction (i.e., relative risk) in DAS28
for biologic experienced patients (i.e., lines 2 and later).}

\item{nma_haq_mean}{Posterior means for HAQ NMA parameters for biologic naive 
patients (i.e., 1st line). Change in HAQ from baseline is modeled using a linear model.}

\item{nma_haq_vcov}{Variance-covariance matrix for HAQ NMA paramters for biologic naive
patients (i.e., 1st line). Change in HAQ from baseline is modeled using a linear model.}

\item{nma_haq_rr_lower}{Lower bound for proportion reduction (i.e., relative risk) in HAQ
for biologic experienced patients (i.e., lines 2 and later).}

\item{nma_haq_rr_upper}{Upper bound for proportion reduction (i.e., relative risk) in HAQ
for biologic experienced patients (i.e., lines 2 and later).}

\item{tx_hist}{Is patient biologic naive or experienced. If naive NMA results for biologic 
naive patients are used during 1st line and biologic experienced NMA results are used
for subsequent lines. If experienced, NMA results for biologic
experienced patients are using during 1st and subsequent lines.}

\item{haq_lprog_tx_mean}{Point estimate of linear yearly HAQ progression rate by treatment.}

\item{haq_lprog_tx_se}{Standard error of linear yearly HAQ progression rate by treatment.}

\item{haq_lcgm_pars}{Parameters of LCGM for HAQ progression.}

\item{eular2haq_mean}{Mean HAQ change by Eular response category.}

\item{eular2haq_se}{Standard error of mean HAQ change by Eular response category.}

\item{acr2haq_mean}{Mean HAQ change by ACR response category.}

\item{acr2haq_se}{Standard error of mean HAQ change by ACR response category.}

\item{acr2sdai_lower}{Lower bound for change in SDAI by ACR response category.}

\item{acr2sdai_upper}{Upper bound for change in SDAI by ACR response category.}

\item{acr2cdai_lower}{Lower bound for change in CDAI by ACR response category.}

\item{acr2cdai_upper}{Upper bound for change in CDAI by ACR response category.}

\item{acr2das28_lower}{Lower bound for change in DAS28 by ACR response category.}

\item{acr2das28_upper}{Upper bound for change in DAS28 by ACR response category.}

\item{haq_lprog_age_mean}{Impact of age on annual linear haq progression rate.}

\item{haq_lprog_age_se}{Standard error of impact of age on annual linear haq progression rate.}

\item{hosp_days_mean}{Vector denoting average number of hospital days for HAQ < 0.5,
0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5.}

\item{hosp_days_se}{Vector denoting standard error of average number of hospital days for HAQ < 0.5,
0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5.}

\item{hosp_cost_mean}{Mean of daily hospital cost.}

\item{hosp_cost_se}{Standard error of dail hospital cost.}

\item{mgmt_cost_mean}{Mean of costs of services (in order: chest x-ray, x-ray visit,
outpatient followup, Mantoux tuberculin skin test) for general management of RA.}

\item{mgmt_cost_se}{Standard error of mean of costs of services for general management of RA
for general management of RA.}

\item{pl_mean}{Mean annual productivity loss per 1-unit increase in HAQ.}

\item{pl_se}{Standard error of mean annual productivity loss per 1-unit increase in HAQ.}

\item{ttsi}{Paramters of survival model used to estimate time to serious infection.}

\item{si_cost}{Cost of a serious infection.}

\item{si_cost_range}{Range used to vary serious infection cost. Default is to calculate upper and lower bound by multiplying 
\code{si_cost} by 1 +/- 0.2 (i.e. a 20\% change).}

\item{si_ul}{One month loss in utility from a serious infection.}

\item{si_ul_range}{Range used to vary serious infection utility loss. Default is to calculate upper and lower bound by multiplying 
\code{si_ul} by 1 +/- 0.2 (i.e. a 20\% change).}

\item{tx_names}{Vector of treatment names.}
}
\value{
List containing samples for the following model parameters:

\describe{
 \item{tx.cost}{\code{data.table} of treatment cost data.}
 \item{ltmale}{Lifetable for males with qx transformed using the logit function.}
  \item{ltfemale}{Lifetable for females with qx transformed using the logit function.}
 \item{acr2eular}{An array of matrices. Each matrix represents a random sample of the conditional 
 probability of each EULAR response category for a given ACR response.}
  \item{logor.mort}{Matrix of log odds ratio used to adjust mortality. One row for each sample
  and one column for each variable used to adjust mortality.}
  \item{mort.loghr.haqdif}{Matrix of the log hazard ratio of the impact of a change in HAQ from baseline on mortality. Columns denote
  hazard ratios at times < 6 months, months 6 - <12, months 12 - <24, months 24 - <36, and months 36+.}
  \item{ttd.eular.mod}{Matrix of coefficients from a survival model. From a survival model for moderate Eular responders}
  \item{ttd.eular.good}{Matrix of coefficients for the location parameter and a vector
   of sampled values of the ancillary parameter. From a survival model for good Eular responders.}
  \item{acr}{A list containing six objects. \code{p1} is an array containing information on the
  probability of achieving the four mutually exclusive ACR categories (ACR<20, ACR 20-50,
  ACR 50-70, ACR70+). The array stores matrices with each row each row a sampled parameter value
  and columns denoting the mutuaully exclusive ACR categories. Therapies are indexed using
  the third dimension. \code{p1.overalap} is identical to \code{p1} but for overlapping ACR
  response categories (ACR <20, ACR 20+, ACR 50+, ACR 70+). \code{p1} and \code{p1.overlap} are
  for the first treatment line. \code{p2} and \code{p2.overlap} are the same as 
  \code{p1} and \code{p1.overlap} but for treatments second line and later. \code{rr} is the 
  sampled values of the probabilility of overalapping ACR response categories for biologic
  experienced patients relative to the probability for biologic naive patients (i.e., 
  the relative risk). Finally, \code{pars} is a matrix of sampled values of model coefficients 
  from the ordered logistic regression used for the NMA.} 
  \item{das28}{A list of four elements. \code{dy1} and \code{dy2} are matrices with each row 
  a sampled parameter value and columns denoting the mean change in DAS28 for each
  therapy. \code{dy1} is for first line treatment and \code{dy2} is for second line treatment
  and later. \code{pars} is a matrix of sampled values of model coefficients from the linear
 model used for the NMA. Finally, \code{rr} is the sampled values of the mean change in
   DAS28 from baseline for biologic experienced patients as a fraction of the mean change for
   biologic naive patients.}
  \item{haq}{Identical to DAS28 but for the HAQ score.}
  \item{eular2haq}{A matrix of sampled HAQ changes by Eular response category. The matrix has
   three columns for no response, moderate response, and good response.}
   \item{acr2haq}{A matrix of sampled HAQ changes by ACR response category. The matrix has
   four columns for ACR < 20, ACR 20-50, ACR 50-70, and ACR 70+.}
  \item{haq.lprog.tx}{A matrix of sampled yearly linear change in HAQ by treatment.
   The matrix has one column for each treatment in \code{iviRA::treatments}.}
   \item{haq.lprog.age}{A matrix of sampled yearly linear change in HAQ by age. The matrix
   has three columns for age < 40, age 40-64, and age 65+.}
   \item{haq.lcgm}{A list of two elements containing parameters from the latent class growth
    model. The first element is \code{delta} which is a an array of sampled matrices with
     each matrix containing coefficients predicting class membership. Rows are classes and columns index
    variables. \code{beta} is similar to \code{delta}, but each matrix contains coefficients 
    predicting HAQ as a function of time using a quadratic polynomial model.
   }
   \item{utility.mixture}{A list containing samples of all parameters in the Hernandez Alva (2013) mixture model. See 'Sampled mixture model parameters'
   for details.}
   \item{utility.wailoo}{A matrix of sampled regression coefficients from the model mapping HAQ to EQ5D utility in Wailoo (2006). Variables are
   (in order) "int" (intercept), "age" (patient age), "dis_dur" (disease duration), "haq0" (baseline HAQ), "male" (1 = male, 0 = female),
   "prev_dmards" (number of previous DMARDs), and "haq" (current HAQ).}
   \item{hosp.cost}{A list of two matrices \code{hosp.days} and \code{cost.pday}. \code{hosp.days} is sample of hospital days by HAQ category; the 
   columns of the matrix are the six HAQ categories (HAQ < 0.5, 0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5). 
   in \code{hosp.days} are HAQ. \code{cost.pday} is a sample of the costs per hospital day by HAQ category; the columns are the same six HAQ
   categories as in \code{hosp.days}.}
   \item{mgmt.cost}{Matrix of sampled values of general management costs. Each column is a different category of costs (
   chest x-ray, x-ray visit, outpatient follow-up, and Mantoux tuberculin skin test). }
   \item{prod.loss}{Vector of sampled values of decrease in wages (e.g. productivity loss) per unit increase in HAQ.}
   \item{ttsi}{Matrix of coefficients from a time-to event model predicting time to serious infection.
    Columns numbers coincide with therapy indices.}
   \item{si.cost}{Vector of sampled values of the medical cost of a serious infection.}
   \item{si.ul}{Vector of the sampled values of the annualized utility loss from a serious infection.}
}
}
\description{
Randomly draw parameters from their (joint) probability distribution.
}
\section{Time to treatment discontinuation}{

Time to treatment discontinuation parameters should be contained in a list of lists. The top-level
 list identifies the name of the probability distribution; the possible distributions are the 
 exponential (\code{exponential}), Weibull (\code{weibull}), Gompertz (\code{gompertz}),
gamma (\code{gamma}), log-logistic (\code{llogis}), lognormal (\code{lnorm}), and generalized gamma (\code{gengamma}). Each distribution 
should also contain a list with five elements:
\describe{
\item{est}{A vector of the maximum likelihood estimates of the parameters. }
\item{vcov}{The variance-covariance of the parameters.}
\item{loc.index}{A vector of the indices of the location parameters.}
\item{anc1.index}{A vector of the indices of the first ancillary parameter.}
\item{anc2.index}{A vector of indices of the second ancillary parameter.}
}
The maximum likelihood estimates should be transformed to the real line. For example, if the model is fit using 
\code{flexsurvreg} in the \code{flexsurv} package, the output should be returned from \code{res.t}.
}

\section{Sampled mixture model parameters}{

The sampled mixture model parameters are contained in a list containing the following:
\describe{
\item{beta1}{Coefficients for class 1 explanatory variables. A matrix of random draws where each
 column is an explanatory variable.}
\item{beta2}{Coefficients for class 2 explanatory variables. A matrix of random draws where each
 column is an explanatory variable.}
\item{beta3}{Coefficients for class 3 explanatory variables. A matrix of random draws where each
 column is an explanatory variable.}
\item{beta4}{Coefficients for class 4 explanatory variables. A matrix of random draws where each
 column is an explanatory variable.}
\item{alpha1}{Random effects intecept term for class 1. A vector of random draws.}
\item{alpha2}{Random effects intecept term for class 2. A vector of random draws.}
\item{alpha3}{Random effects intecept term for class 3. A vector of random draws.}
\item{alpha4}{Random effects intecept term for class 4. A vector of random draws.}
\item{alpha}{Random effects term for male indicator variable. A vector of random draws.}
\item{epsilon1}{Variance for class 1. A vector of random draws.}
\item{epsilon2}{Variance for class 2. A vector of random draws.}
\item{epsilon3}{Variance for class 3. A vector of random draws.}
\item{epsilon4}{Variance for class 4. A vector of random draws.}
\item{mu}{Random effects variance term. A vector of random draws.}
\item{delta}{Coefficients for explanatory variables explaining the probbaility
of class membership. An array of matrices where each matrix is a random draw. There are three rows in each matrix (one for each class) and
four columns (one for each explanatory variable).}
}
The list also contains summary statistics for pain and HAQ, which is needed to simulate pain. 
In particular, the object 'pain' in the list is a list containing:
\describe{
\item{pain.mean}{Mean of pain score in the population.}
\item{haq.mean}{Mean of HAQ score in the population.}
\item{pain.var}{Variance of pain score in the population.}
\item{haq.var}{Variance of HAQ in the population.}
\item{painhaq.cor}{Correlation between pain and HAQ in the population.}
}
}

