% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pars.R
\name{sample_pars_utility_mixture}
\alias{sample_pars_utility_mixture}
\title{Sample parameters for mixture model utility mapping}
\usage{
sample_pars_utility_mixture(nsims, pain)
}
\value{
List containing posterior sample of paramaters from Hernandez Alva (2013) 
mixture model
}
\description{
Sample parameters for mixture model utility mapping.
}
