% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_iviRA}
\alias{sim_iviRA}
\title{The IVI-RA simulation}
\usage{
sim_iviRA(arms, input_data, pars, model_structure, ttd_dist = "lnorm",
  si_dist = "exp", max_months = NULL,
  cdmards_ind = which(therapy.pars$info$sname == "cdmards"),
  nbt_ind = which(therapy.pars$info$sname == "nbt"), check = TRUE)
}
\arguments{
\item{arms}{Name of arms in the treatment treatment sequence. May be a vector consisting of 
a single treatment sequence or a matrix of unique sequences for each patient.}

\item{input_data}{An object of class 'input_data' returned from \link{get_input_data}.}

\item{pars}{List of sampled parameter values generated from \link{sample_pars}.}

\item{model_structure}{Object of class model structure generated from \link{sample_pars}.}
}
\description{
Run the IVI-RA individual patient simulation model.
}
