% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_iviRA}
\alias{sim_iviRA}
\title{The IVI-RA simulation}
\usage{
sim_iviRA(arms, input_data, pars, model_structure, max_months = NULL,
  treatment_lookup = iviRA::treatments$sname,
  cdmards_ind = which(therapy.pars$info$sname == "cdmards"),
  nbt_ind = which(therapy.pars$info$sname == "nbt"), check = TRUE)
}
\arguments{
\item{arms}{Name of arms in the treatment treatment sequence. May be a vector consisting of 
a single treatment sequence or a matrix of unique sequences for each patient.}

\item{input_data}{An object of class 'input_data' returned from \link{get_input_data}.}

\item{pars}{List of sampled parameter values generated from \link{sample_pars}.}

\item{model_structure}{Object of class model structure generated from \link{sample_pars}.}

\item{max_months}{Maximum number of months to run the model for. Default is NULL which implies that
the model is simulated over each patient's lifetime.}

\item{cdmards_ind}{Index for cDMARDs.}

\item{nbt_ind}{Index for the non-biologic (i.e. a term used to define a selection of treatments clinicians use
after the last biologic in a treatment sequence).}

\item{treatments_lookup}{Vector of names of all treatments included in the parameter
estimates. Index of of treatments in \code{arms} are matched against treatments in
\code{treatments_lookup} by name. Indices of treatment-specific parameter estimates must be 
in the same order as treatments in \code{treatments_lookup}.}
}
\description{
Run the IVI-RA individual patient simulation model.
}
