% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_iviRA}
\alias{sim_iviRA}
\title{The IVI-RA simulation}
\usage{
sim_iviRA(arms, input_data, pars, model_structures, max_months = NULL,
  tx_data = iviRA::treatments, hist = c("naive", "experienced"),
  output = c("data", "summary"), discount_qalys = 0.03,
  discount_cost = 0.03, check = TRUE)
}
\arguments{
\item{arms}{Name of arms in the treatment treatment sequence. May be a vector consisting of 
a single treatment sequence or a matrix of unique sequences for each patient.}

\item{input_data}{An object of class 'input_data' returned from \link{get_input_data}.}

\item{pars}{List of sampled parameter values generated from \link{sample_pars}.}

\item{max_months}{Maximum number of months to run the model for. Default is NULL which implies that
the model is simulated over each patient's lifetime.}

\item{tx_data}{Dataset of treatments with columns names equivalent to \code{iviRA::treatments}.
Index of of treatments in \code{arms} are matched against treatments in
\code{tx_data$sname} by name. Indices of treatment-specific parameter estimates must be 
in the same order as treatments in \code{tx_data}.}

\item{hist}{Is the patient biologic naive or biologic experienced?}

\item{output}{Specifies the format of output returned from the simulation. Options are \code{data} 
and \code{summary}. When \code{data} is specified, each simulated value (i.e, by model,
sampled parameter set, individual, and time-period) is returned in a \code{data.table}. If 
\code{summary} is selected, then only summary measures are returned.}

\item{discount_qalys}{Discount rate for QALYs. Only used when \code{output = "summary"}; 
otherwise, discounts can be applied to the simulated output.}

\item{discount_cost}{Discount rate for cost variables. Only used when \code{output = "summary"};
otherwise, discounts can be applied to the simulated output.}

\item{model_structure}{Object of class model structure generated from \link{select_model_structure}.}
}
\value{
The \code{output = "data"} options returns all simulated output. However, since output is 
returned for each model, sampled parameter set, individual, and model cycle, 
the size of the output can be very large and can cause memory management issues. The 
\code{output = "summary"}, which only provides summaries of the simulation output, can be 
useful in the cases. 

When \code{output = "data"} is selected, the simulation returns a \code{data.table} with the 
following columns:
\describe{
\item{sim}{Simulation number. Indexed from 1 to S where S is the number of randomly sampled parameter sets (e.g. n from \link{sample_pars}).}
\item{id}{ID number. Indexed from 1 to N where N is the number of simulated patients (e.g. n from \link{sample_pats}).}
\item{month}{Month since a simulated patient began the first treatment in a treatment sequence.}
\item{tx}{Treatment used. Given J total therapies, the first J - 1 therapies match the indices from \code{arminds}. The final \code{tx}
 is always the non-biologic treatment (\code{nbt_ind}).}
\item{tx_seq}{Number of treatment in a treatment sequence. First treatment equal to 1, 
second treatment equal to 2, \ldots}
\item{tx_cycle}{Number of model cycles since a patient began taking a given treatment in a treatment sequence. \code{tx_cycle} = 1
during the initial 6-month treatment period.}
\item{death}{Equal to 1 if patient died during the model cycle and 0 otherwise. }
\item{age}{Age of patient which increases with the model cycles.}
\item{ttd}{Time to treatment discontinuation. Measured in terms of model cycles (e.g. ttd = 2 if treatment will discontinue in 
1 year given 6-months cycles). \code{ttd} is measured at the end of each cycle. Patients switch treatments during the cycle in which \code{ttd} 
becomes negative and HAQ rebounds during the cycle.}
\item{acr}{Simulated ACR response during the initial 6-month period for a new treatment Constant within \code{tx}.
Categories are 0 (ACR < 20), 1 (ACR 20-50), 2 (ACR 50-70), and 3 (ACR 70+).}
\item{eular}{Simulated EULAR response during the initial 6-month period for a new treatment Constant within \code{tx}. 
Categories are 0 (no EULAR response), 1 (moderate EULAR response), and 2 (good EULAR response).}
\item{das28}{Disease Activity Score 28 (DAS28).}
\item{sdai}{Simplified Disease Activity Index (SDAI).}
\item{cdai}{Clinical Disease Activity Index (CDAI).}
\item{haq}{HAQ score. Restricted to range between 0 and 3.}
\item{ttsi}{Time to serious infection. Like \code{ttd}, measured in terms of model cycles. \code{ttsi} is measured at the end of each 
cycle.}
\item{si}{Equal to 1 if treatment discontinuation was caused by a serious infection and 0 otherwise.}
\item{yrlen}{Length of a model cycle in years. Equal to 0.5 given 6-month cycles.}
\item{tx_cost}{Treatment costs after discounts and rebates.}
\item{hosp_cost}{Hospitalization costs.}
\item{mgmt_cost}{General management costs.}
\item{si_cost}{Cost due to serious infections.}
\item{prod_loss}{Productivity loss (i.e., lost earnings).}
\item{utility}{Simulated utility score.}
\item{qalys}{Quality-adjusted life-years (QALYs).}
}
}
\description{
Run the IVI-RA individual patient simulation model.
}
