% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{sim_utility_wailoo}
\alias{sim_utility_wailoo}
\title{Simulate utility from simulated HAQ score using Wailoo (2006) model}
\usage{
sim_utility_wailoo(simhaq, haq0, male, prev_dmards, coefs, check = TRUE)
}
\arguments{
\item{simhaq}{Simulation output from \link{sim_haq}. Variables needed are \code{sim}, \code{id},
\code{age}, and \code{haq}.}

\item{haq0}{HAQ score at baseline.}

\item{male}{Indicator = 1 for males and 0 for females.}

\item{prev_dmards}{Number of previous DMARDs.}

\item{coefs}{Matrix of coefficients needed to simulate utility using the Wailoo (2006) model. 
See the documentation in \link{sample_pars} for details. Note that the matrix columns must be in 
the same order as generated by \link{sample_pars}.}

\item{check}{Should the function check parameters and input data passed to the model? Default is TRUE.}
}
\value{
Vector of utility scores.
}
\description{
Simulate utility from HAQ score simulated using \code{sim_haq} using model from
Wailoo (2006).
}
\details{
Note that disease duration is set to 18.65 years, which
is the mean value from the Wailoo (2006) paper used for the parameter estimates. Age and 
the HAQ score are taken from the simulation output.
}
