% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treat.switch}
\alias{treat.switch}
\title{Treatment switching at 6 months}
\format{A data.table containing the following columns:
\describe{
  \item{disease_activity}{Name of variable. Includes the intercept, an indicator variable for whether
  disease activity is moderate (CDAI 10-22), and an indicator variable for 
  whether disease activity is high (CDAI greater than 22).}
  \item{p}{Probability of switching treatment. }
  \item{p_se}{Standard error of probability of switching treatment.}
  \item{alpha}{Alpha parameter in beta distribution.}
  \item{beta}{Beta parameter in a beta distribution.}
}}
\source{
Zhang, Jie, et al. "Thresholds in disease activity for switching biologics in rheumatoid 
arthritis patients: experience from a large US cohort." 
Arthritis care & research 63.12 (2011): 1672-1679.
}
\usage{
treat.switch
}
\description{
Probability of treatment switching at 6 months by disease activity level (remission, low,
moderate, high)
}
\keyword{datasets}
