% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttd}
\alias{ttd}
\alias{ttd.eular}
\alias{ttd.da}
\alias{ttd.all}
\title{Time to treatment discontinuation parameters}
\format{A list of lists of lists. The top level list is the level of EULAR response or
 disease activity. The second level list is the parametric distribution used. 
 The third level list contains the parameters of a given parametetric fit as described in
 \link{sample_pars}.}
\source{
Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, 
golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously
 treated with disease-modifying antirheumatic drugs and after the failure of conventional 
 disease-modifying antirheumatic drugs only: systematic review and economic evaluation." 
Health Technology Assessment 20.35 (2016): 1-610.

Zhang, Jie, et al. "Thresholds in disease activity for switching biologics in rheumatoid 
arthritis patients: experience from a large US cohort." 
Arthritis care & research 63.12 (2011): 1672-1679.
}
\usage{
ttd.eular

ttd.da

ttd.all
}
\description{
\code{ttd.eular} and \code{ttd.da} are lists of lists containing treatment discontinuation
parameters stratified by EULAR response (moderate, high) and disease activity 
(remission, low, moderate, high) respectively; \code{ttd.all} is a list of lists containing
unstratified (i.e., all patients) treatment discontinuation parameters.
}
\details{
Time to treatment discontinuation paramters for each level of EULAR response or disease activity
are represented with a list as described in \link{sample_pars}. Models were fit using \emph{flexsurv}.
Survival curves are based on analyses of reconstructed individual patient data.
}
\keyword{datasets}
